Update local server port and enhance security configuration
Changed the port for local server in OpenAPI configuration from 8088 to 8080 and enhanced the security configurations by implementing a white-list method for JWT filtering. This update provides a more structured and secure way to handle requests and responses for specific endpoints in the application. Made respective changes in "JwtFilter" and "SecurityConfig" Java files.